Cleaned up the namese and set a more predictable return value.
When getting a URL from CRAN or a custom repo using the `url_from_cran()` or `url_from_custom_repo()` functions respectively, you'll always get the result passed from `parse_urls()`. And now `parse_urls()` can handle empty values, the return value is the same for empty character string and NA values - both of which could be returned from the dataframe check depending on whether the package and/or URL is found.